article-extraction/requirements.txt
requests
nltk==3.3
newspaper3k==0.2.6
